Move macOS sandbox builders into codex-sandboxing
