+++ codex-rs/sandboxing/Cargo.toml
@@ -12,8 +12,15 @@ path = "src/lib.rs"
 workspace = true
 
 [dependencies]
+codex-network-proxy = { workspace = true }
 codex-protocol = { workspace = true }
+codex-utils-absolute-path = { workspace = true }
+dirs = { workspace = true }
+libc = { workspace = true }
 serde_json = { workspace = true }
+tracing = { workspace = true, features = ["log"] }
+url = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
+tempfile = { workspace = true }
