+++ codex-rs/sandboxing/src/lib.rs
@@ -1,1 +1,6 @@
 pub mod landlock;
+pub mod macos_permissions;
+#[cfg(target_os = "macos")]
+pub mod seatbelt;
+#[cfg(target_os = "macos")]
+mod seatbelt_permissions;
