+++ codex-rs/sandboxing/src/macos_permissions_tests.rs
