+++ codex-rs/sandboxing/src/restricted_read_only_platform_defaults.sbpl
